requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
